Updated variable type of 'size' field in 'sata_dev' struct from int to unsigned long long.

+++ sata2_driver_v1_00_a/sata_drv.h
@@ -30,7 +30,7 @@
  * The internal representation of our device.
  */
 struct sata_dev {
-        int size;                       /* Device size in sectors */
+        unsigned long long size;        /* Device size in sectors */
         short users;                    /* How many users */
         short media_change;             /* Flag a media change? */
         spinlock_t lock;                /* For mutual exclusion */
